+++ src/config/firefox-desktop.js
@@ -51,6 +51,9 @@ export default {
     if (pr.record_in_processes[0] === 'all') {
       pr.record_in_processes = ['main', 'content', 'gpu'];
     }
+    if (pr.record_in_processes[0] === 'all_childs') {
+      pr.record_in_processes = ['content', 'gpu'];
+    }
     return pr;
   },
   setDefaultsForProbe(store, probe) {