+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Select prerelease channel if current channel = release & probe only exists in
   prerelease ([#492](https://github.com/mozilla/glam/pull/492))
+- Change GLAM API response to be more tabular
+  ([#494](https://github.com/mozilla/glam/pull/494))
 
 ## [2020.5.0](https://github.com/mozilla/glam/compare/2020.4.2...2020.5.0) (2020-05-12)